Split hal and lib acpi_tables
